chore: move PreferenceImpl to its own class
